Corregidos controladores de un plan estratégico para que se tengan en cuenta los ids de planes no existentes y se lance el respectivo error.

+++ icasus/app_code/plan_borrar.php
@@ -18,20 +18,29 @@ if (filter_has_var(INPUT_GET, 'id_plan') && filter_has_var(INPUT_GET, 'id_entida
     $id_plan = filter_input(INPUT_GET, 'id_plan', FILTER_SANITIZE_NUMBER_INT);
     $id_entidad = filter_input(INPUT_GET, 'id_entidad', FILTER_SANITIZE_NUMBER_INT);
 
+    //Obtenemos los datos del plan
     $plan = new Plan();
-    $plan->load("id = $id_plan");
-    $linea = new Linea();
-    $lineas_plan = $linea->Find("id_plan=$id_plan");
-    if (!$lineas_plan && $control)
+    if ($plan->load("id = $id_plan"))
     {
-        $exito = MSG_PLAN_BORRADO . " " . $plan->anyo_inicio . "-" . ($plan->anyo_inicio + $plan->duracion - 1);
-        $plan->delete();
-        header("Location: index.php?page=entidad_mostrar&id_entidad=$id_entidad&exito=$exito");
+        $smarty->assign('plan', $plan);
+        $linea = new Linea();
+        $lineas_plan = $linea->Find("id_plan=$id_plan");
+        if (!$lineas_plan && $control)
+        {
+            $exito = MSG_PLAN_BORRADO . " " . $plan->anyo_inicio . "-" . ($plan->anyo_inicio + $plan->duracion - 1);
+            $plan->delete();
+            header("Location: index.php?page=entidad_mostrar&id_entidad=$id_entidad&exito=$exito");
+        }
+        else
+        {
+            $aviso = MSG_PLAN_BORRAR;
+            header("Location: index.php?page=plan_mostrar&id_plan=$id_plan&id_entidad=$id_entidad&aviso=$aviso");
+        }
     }
     else
     {
-        $aviso = MSG_PLAN_BORRAR;
-        header("Location: index.php?page=plan_mostrar&id_plan=$id_plan&id_entidad=$id_entidad&aviso=$aviso");
+        $error = ERR_PLAN_MOSTRAR;
+        header("location:index.php?page=error&error=$error");
     }
 }
 else
